+++ aiworkspace-20210204/core/Models/ValidateConnectionRequest.cs
@@ -1,0 +1,56 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.AIWorkSpace20210204.Models
+{
+    public class ValidateConnectionRequest : TeaModel {
+        [NameInMap("Configs")]
+        [Validation(Required=false)]
+        public Dictionary<string, string> Configs { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>conn-x2wz5zvvbyf2420fi9</para>
+        /// </summary>
+        [NameInMap("ConnectionId")]
+        [Validation(Required=false)]
+        public string ConnectionId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>DashScopeConnection</para>
+        /// </summary>
+        [NameInMap("ConnectionType")]
+        [Validation(Required=false)]
+        public string ConnectionType { get; set; }
+
+        [NameInMap("Secrets")]
+        [Validation(Required=false)]
+        public Dictionary<string, string> Secrets { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>Connectivity</para>
+        /// </summary>
+        [NameInMap("ValidateType")]
+        [Validation(Required=false)]
+        public string ValidateType { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>1234</para>
+        /// </summary>
+        [NameInMap("WorkspaceId")]
+        [Validation(Required=false)]
+        public string WorkspaceId { get; set; }
+
+    }
+
+}