+++ aiworkspace-20210204/core/Models/ValidateConnectionResponseBody.cs
@@ -1,0 +1,50 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.AIWorkSpace20210204.Models
+{
+    public class ValidateConnectionResponseBody : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>403</para>
+        /// </summary>
+        [NameInMap("Code")]
+        [Validation(Required=false)]
+        public string Code { get; set; }
+
+        [NameInMap("Details")]
+        [Validation(Required=false)]
+        public Dictionary<string, string> Details { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>The connection is reachable.</para>
+        /// </summary>
+        [NameInMap("Message")]
+        [Validation(Required=false)]
+        public string Message { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>5A14FA81-DD4E-******-6343FE44B941</para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>success</para>
+        /// </summary>
+        [NameInMap("Status")]
+        [Validation(Required=false)]
+        public string Status { get; set; }
+
+    }
+
+}